Resolve dual BaseEvent definitions and standardize event system
EVENT SYSTEM CLEANUP:
- Eliminated duplicate BaseEvent class definition in schemas.py
- Migrated all event schemas to use proper BaseEvent from base.py
- Converted all events to use modern dataclass pattern
- Added @register_event decorators for automatic event registration

ARCHITECTURAL IMPROVEMENTS:
- All events now inherit from single, comprehensive BaseEvent
- Proper event metadata management with event_id, timestamp, version
- Consistent event payload structure across all event types
- Automatic event type registration for better type safety

MODERNIZATION:
- Replaced manual __init__ with dataclass field definitions
- Better type hints and optional parameter handling
- Consistent payload serialization across all events
- Improved maintainability and extensibility

EVENTS MIGRATED:
- ClientCreatedEvent, ClientUpdatedEvent
- TaskCreatedEvent, TaskUpdatedEvent, TaskStatusChangedEvent, TaskDeletedEvent
- DocumentAnalysisStartedEvent, DocumentAnalysisCompletedEvent, DocumentAnalysisFailedEvent
- DocumentCreatedEvent, DocumentUpdatedEvent
- ProjectCreatedEvent, ProjectUpdatedEvent
- ErrorEvent

This resolves the dual BaseEvent definitions issue from the architectural review.
